+++ changelog.md
@@ -93,7 +93,10 @@ Proof-of-concept version.
 * CMake force include are platform independent
 * Added Framework and FrameworkDir
 * temporary output files are under <output_dir> / <recipe> / <md5 of dir> / <rel>
+* Comparing chai recipes and books
+* [Project|Output|Temporary] dir are exposed in chai (read-only)
 * [wip] Support for HTML output
+
 
 ## Next